% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/gtf_tools.R
\name{gtf_reader}
\alias{gtf_reader}
\title{Reads GTFs}
\usage{
gtf_reader(filename)
}
\arguments{
\item{filename}{Path to gtf file.}
}
\description{
Does what is says.
}
\examples{
gtf_reader()
}
\keyword{GTF}

