% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/gtf_tools.R
\name{gtf_writer}
\alias{gtf_writer}
\title{Writes GTFs}
\usage{
gtf_writer(x, filename)
}
\arguments{
\item{x}{GTF object.}

\item{filename}{new gtf filename.}
}
\description{
Does what is says.
}
\examples{
gtf_writer()
}
\keyword{GTF}

