% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/gtf_tools.R
\name{remove_unstranded}
\alias{remove_unstranded}
\title{Removes unstranded from gtf read into R.}
\usage{
remove_unstranded(x)
}
\arguments{
\item{x}{GTF object.}
}
\description{
Assumes stranded info is in column 7.
}
\examples{
remove_unstranded(mtcars)
}
\keyword{GTF}

