% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/gtf_tools.R
\name{replace_id_str}
\alias{replace_id_str}
\title{Update id in plink ped}
\usage{
replace_id_str(x, list)
}
\arguments{
\item{x}{Character vector of lines from ped}

\item{List}{two column text file with pattern and replacement}
}
\description{
Update Ids in plink ped file based on two column data frame
}
\examples{
replace_id_str()
}
\keyword{plink}

