% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/gtf_tools.R
\name{replace_id_str}
\alias{replace_id_str}
\title{Update id in vector of IDs, eg. plink fam.}
\usage{
replace_id_str(x, list)
}
\arguments{
\item{x}{Character vector of IDs to update.}

\item{List}{two column dataframe with pattern and replacement}
}
\description{
Update Ids in dataframe column based on two column data frame
}
\examples{
delete <- apply(HDanimals, 2, replace_id_str, list = id_changes_complete)
}
\keyword{plink}

